+++ CHANGES.md
@@ -1,5 +1,10 @@
 # Changes
 
+In next release ...
+
+- Added browser events `AnimationEnd`, `AnimationIteration`, and
+  `AnimationStart`.
+
 1.0.1 (2019-01-13)
 ------------------